refactor(box): implement common rate limit middleware
