+++ apps/box/src/app/auth/service.test.ts
@@ -5,7 +5,7 @@ import { db } from '@/database/client';
 import { organisationsTable } from '@/database/schema';
 import { inngest } from '@/inngest/client';
 import { decrypt } from '@/common/crypto';
-import { BoxError } from '@/connectors/common/error';
+import { BoxError } from '@/connectors/box/common/error';
 import { setupOrganisation } from './service';
 
 const code = 'some-code';
