+++ apps/box/src/connectors/box/auth.test.ts
@@ -2,7 +2,7 @@ import { http } from 'msw';
 import { describe, expect, test, beforeEach } from 'vitest';
 import { server } from '@elba-security/test-utils';
 import { env } from '@/common/env';
-import { BoxError } from '../common/error';
+import { BoxError } from './common/error';
 import { getRefreshToken, getToken } from './auth';
 
 const validCode = '1234';
