--- apps/box/src/inngest/middlewares/rate-limit-middleware.ts
@@ -1,42 +1,0 @@
-import { InngestMiddleware, RetryAfterError } from 'inngest';
-import { BoxError } from '@/connectors/common/error';
-
-export const rateLimitMiddleware = new InngestMiddleware({
-  name: 'rate-limit',
-  init: () => {
-    return {
-      onFunctionRun: ({ fn }) => {
-        return {
-          transformOutput: (ctx) => {
-            const {
-              result: { error, ...result },
-              ...context
-            } = ctx;
-
-            if (!(error instanceof BoxError)) {
-              return;
-            }
-
-            if (error.response?.status === 429 && error.response.headers.get('retry-after')) {
-              const retryAfter = error.response.headers.get('retry-after') || 60;
-
-              return {
-                ...context,
-                result: {
-                  ...result,
-                  error: new RetryAfterError(
-                    `API rate limit reached by '${fn.name}', retry after ${retryAfter} seconds.`,
-                    `${retryAfter}s`,
-                    {
-                      cause: error,
-                    }
-                  ),
-                },
-              };
-            }
-          },
-        };
-      },
-    };
-  },
-});
